+++ Deveroom.VisualStudio/ProjectSystem/Actions/QuestionDescription.cs
@@ -9,13 +9,15 @@ namespace Deveroom.VisualStudio.ProjectSystem.Actions
         public bool IncludeCancel { get; }
         public Action<QuestionDescription> YesCommand { get; }
         public Action<QuestionDescription> NoCommand { get; }
+        public bool NoCommandIsDefault { get; }
 
-        public QuestionDescription(string title, string description, Action<QuestionDescription> yesCommand, Action<QuestionDescription> noCommand = null, bool includeCancel = false)
+        public QuestionDescription(string title, string description, Action<QuestionDescription> yesCommand, Action<QuestionDescription> noCommand = null, bool includeCancel = false, bool noCommandIsDefault = false)
         {
             Description = description;
             YesCommand = yesCommand;
             NoCommand = noCommand;
             IncludeCancel = includeCancel;
+            NoCommandIsDefault = noCommandIsDefault;
             Title = title;
         }
     }